move: 포인트 공통 컴포넌트 shared 폴더로 이동 (#111)
