+++ apps/web/src/features/user/mypage/components/main/MyPageUserProfile.tsx
@@ -7,7 +7,7 @@ import AddressStatus from 'src/features/user/mypage/components/main/AddressStatu
 import BaseAvatar from 'src/shared/ui/BaseAvatar';
 import BaseBadge from 'src/shared/ui/BaseBadge';
 import BaseCard from 'src/widgets/BaseCard';
-import PointDisplay from 'src/widgets/PointDisplay';
+import PointDisplay from 'src/shared/ui/PointDisplay';
 import type { RoleType, UserMetadata } from 'src/entities/user/mypage/main/types';
 
 const MyPageUserProfile = ({ data }: UserMetadata) => {
