+++ apps/web/src/shared/ui/PointDisplay.tsx
